+++ aasrp/tests/test_templatetags.py
@@ -4,15 +4,12 @@ Tests for the template tags
 
 # Django
 from django.template import Context, Template
-from django.test import TestCase, override_settings
+from django.test import TestCase
 
 # Alliance Auth
 from allianceauth.tests.auth_utils import AuthUtils
 
 # AA SRP
-from aasrp import __version__
-from aasrp.constants import PACKAGE_NAME
-from aasrp.helper.static_files import calculate_integrity_hash
 from aasrp.models import get_sentinel_user
 from aasrp.tests.utils import create_fake_user
 
@@ -602,95 +599,3 @@ class TestMainAllianceId(TestCase):
 
         # then
         self.assertEqual(first=result, second="1")
-
-
-class TestVersionedStatic(TestCase):
-    """
-    Tests for aasrp_static template tag
-    """
-
-    @override_settings(DEBUG=False)
-    def test_versioned_static(self):
-        """
-        Test should return the versioned static
-
-        :return:
-        :rtype:
-        """
-
-        context = Context(dict_={"version": __version__})
-        template_to_render = Template(
-            template_string=(
-                "{% load aasrp %}"
-                "{% aasrp_static 'css/aa-srp.min.css' %}"
-                "{% aasrp_static 'javascript/aa-srp.min.js' %}"
-            )
-        )
-
-        rendered_template = template_to_render.render(context=context)
-
-        expected_static_css_src = (
-            f'/static/{PACKAGE_NAME}/css/aa-srp.min.css?v={context["version"]}'
-        )
-        expected_static_css_src_integrity = calculate_integrity_hash(
-            "css/aa-srp.min.css"
-        )
-        expected_static_js_src = (
-            f'/static/{PACKAGE_NAME}/javascript/aa-srp.min.js?v={context["version"]}'
-        )
-        expected_static_js_src_integrity = calculate_integrity_hash(
-            "javascript/aa-srp.min.js"
-        )
-
-        self.assertIn(member=expected_static_css_src, container=rendered_template)
-        self.assertIn(
-            member=expected_static_css_src_integrity, container=rendered_template
-        )
-        self.assertIn(member=expected_static_js_src, container=rendered_template)
-        self.assertIn(
-            member=expected_static_js_src_integrity, container=rendered_template
-        )
-
-    @override_settings(DEBUG=True)
-    def test_versioned_static_with_debug_enabled(self) -> None:
-        """
-        Test versioned static template tag with DEBUG enabled
-
-        :return:
-        :rtype:
-        """
-
-        context = Context({"version": __version__})
-        template_to_render = Template(
-            template_string=(
-                "{% load aasrp %}" "{% aasrp_static 'css/aa-srp.min.css' %}"
-            )
-        )
-
-        rendered_template = template_to_render.render(context=context)
-
-        expected_static_css_src = (
-            f'/static/{PACKAGE_NAME}/css/aa-srp.min.css?v={context["version"]}'
-        )
-
-        self.assertIn(member=expected_static_css_src, container=rendered_template)
-        self.assertNotIn(member="integrity=", container=rendered_template)
-
-    @override_settings(DEBUG=False)
-    def test_invalid_file_type(self) -> None:
-        """
-        Test should raise a ValueError for an invalid file type
-
-        :return:
-        :rtype:
-        """
-
-        context = Context({"version": __version__})
-        template_to_render = Template(
-            template_string=(
-                "{% load aasrp %}" "{% aasrp_static 'invalid/invalid.txt' %}"
-            )
-        )
-
-        with self.assertRaises(ValueError):
-            template_to_render.render(context=context)